Create an API to get the list of motorbikes and bicycles
Create an API to get the list of motorbikes and bicycles belonging to a provider, and return the data in the following format:

[   {     "thumb": "Vehicle Name",     "vehicle": [       {         // vehicle-specific details: image, license plate, vehicle ID, etc.       }     ],     "vehicleNumber": 3   },   {     "thumb": "Another Vehicle Name",     "vehicle": [       {         // vehicle-specific details       }     ]   } ]

thumb is the name of the vehicle (e.g., the model or brand).

Each thumb group contains:

A list of vehicles (vehicle) with their individual details such as image, license plate, and ID.

A count of how many vehicles there are (vehicleNumber, optional).

The vehicle creation function will update the number_vehicle field of vehicles that share the same name and owner when another vehicle of the same model is added. Vehicles of the same type must have identical information, except for the following fields: id, vehicle_images, license_plate, and total_rating.

The update function will be separated into two methods:

One for updating shared (common) information.

One for updating individual (specific) information.